requirements.txt
Django==1.8.1
djangorestframework==3.1.1
